+++ MatrixSDK/MXError.h
@@ -33,6 +33,9 @@ FOUNDATION_EXPORT NSString *const kMXErrCodeStringLimitExceeded;
 FOUNDATION_EXPORT NSString *const kMXErrCodeStringUserInUse;
 FOUNDATION_EXPORT NSString *const kMXErrCodeStringRoomInUse;
 FOUNDATION_EXPORT NSString *const kMXErrCodeStringBadPagination;
+FOUNDATION_EXPORT NSString *const kMXErrCodeStringUnauthorized;
+FOUNDATION_EXPORT NSString *const kMXErrCodeStringLoginEmailURLNotYet;
+FOUNDATION_EXPORT NSString *const kMXErrCodeStringThreePIDAuthFailed;
 
 FOUNDATION_EXPORT NSString *const kMXErrorStringInvalidToken;