+++ MatrixSDK/MXSession.m
@@ -28,7 +28,7 @@
 
 #pragma mark - Constants definitions
 
-const NSString *MatrixSDKVersion = @"0.6.4";
+const NSString *MatrixSDKVersion = @"0.6.5";
 NSString *const kMXSessionStateDidChangeNotification = @"kMXSessionStateDidChangeNotification";
 NSString *const kMXSessionNewRoomNotification = @"kMXSessionNewRoomNotification";
 NSString *const kMXSessionWillLeaveRoomNotification = @"kMXSessionWillLeaveRoomNotification";
@@ -934,6 +934,24 @@ typedef void (^MXOnResumeDone)();
     {
         return nil;
     }
+}
+
+- (MXRoom *)roomWithAlias:(NSString *)alias
+{
+    MXRoom *theRoom;
+
+    if (alias)
+    {
+        for (MXRoom *room in rooms.allValues)
+        {
+            if (room.state.aliases && NSNotFound != [room.state.aliases indexOfObject:alias])
+            {
+                theRoom = room;
+                break;
+            }
+        }
+    }
+    return theRoom;
 }
 
 - (NSArray *)rooms